Trait actix_governor::KeyExtractor

pub trait KeyExtractor: Clone + Copy {
    type Key: Clone + Hash + Eq;
    type KeyExtractionError: Display;
    fn extract(
        &self, 
        req: &ServiceRequest
    ) -> Result<Self::Key, Self::KeyExtractionError>;
}

Generic structure of what is needed to extract a rate-limiting key from an incoming request.

Associated Types

type Key: Clone + Hash + Eq

The type of the key.

type KeyExtractionError: Display

The type of the error that can occur if key extraction from the request fails.

Required methods

fn extract(
&self,
req: &ServiceRequest
) -> Result<Self::Key, Self::KeyExtractionError>

Extraction method